simplify public interface
