+++ charm/src/webhook_redelivery/redelivery.py
@@ -3,10 +3,44 @@
 
 """Implementation of the webhook redelivery logic."""
 
+
+from dataclasses import dataclass
 from datetime import datetime
 
-from github import Github
-from webhook_redelivery import WebhookAddress
+
+@dataclass
+class GithubAppAuthDetails:
+    """The details to authenticate with Github using a Github App.
+
+    Attributes:
+        app_id: The Github App (or client) ID.
+        installation_id: The installation ID of the Github App.
+        private_key: The private key to authenticate with Github.
+    """
+
+    app_id: str
+    installation_id: str
+    private_key: str
+
+
+GithubToken = str
+GithubAuthDetails = GithubAppAuthDetails | GithubToken
+
+
+@dataclass
+class WebhookAddress:
+    """The address details to identify the webhook.
+
+    Attributes:
+        github_org: Github organisation where the webhook is registered.
+        github_repo: Github repository, where the webhook is registered. Only applicable for
+            repository webhooks.
+        id: The identifier of the webhook.
+    """
+
+    github_org: str
+    github_repo: str | None
+    id: str
 
 
 class RedeliveryError(Exception):
@@ -14,12 +48,12 @@ class RedeliveryError(Exception):
 
 
 def redeliver_failed_webhooks(
-    github_client: Github, webhook_address: WebhookAddress, since: datetime
+    github_auth: GithubAuthDetails, webhook_address: WebhookAddress, since: datetime
 ) -> int:
     """Redeliver failed webhooks since a given time.
 
     Args:
-        github_client: The GitHub client used for interactions with Github.
+        github_auth: The GitHub authentication details used to interact with the Github API.
         webhook_address: The data to identify the webhook.
         since: The time to redeliver failed webhooks since.
